뷁

+++ TEST/bingsu.cpp
@@ -1,21 +1,21 @@
 #include <bits/stdc++.h>
 using namespace std;
 typedef long long ll;
-int n;
+ll n;
 
 int main () {
     cin >> n;
     priority_queue<ll, vector<ll>, greater<ll> > q;
     for (int i = 0; i < n; ++i) {
-		int a;
+		ll a;
         cin >> a;
         q.push(a);
     } 
-    int res = 0;
+    ll res = 0;
     while (!(q.empty())) {
-        int a = q.top(); q.pop();
-        if (q.empty()) break;
-        int b = q.top(); q.pop();
+        ll a = q.top(); q.pop();
+        if(q.empty()) break;
+        ll b = q.top(); q.pop();
         res += (a+b);
         q.push(a+b);
     }
